+++ changelog.md
@@ -1,5 +1,5 @@
 # 0.2.0 
 ### Update UI, add more function: 
-- Continute a Tomato, continute working on subTask which haven't done yet.
+- Continue a Tomato, continue working on subTask which haven't done yet.
 # 0.1.0
 First commit, some base function